+++ projects/sparse_rl/sparse_dqn.py
@@ -22,8 +22,6 @@
 from __future__ import print_function
 
 import argparse
-import collections
-import copy
 import csv
 import os
 import random
@@ -31,12 +29,11 @@ import random
 import numpy as np
 import torch
 import torch.nn as nn
-import torch.nn.functional as F
-import torch.optim as optim
 import tqdm
 
 from htmresearch.frameworks.pytorch.sparse_net import SparseNet
-from htmresearch.support.openai_utils import create_atari_environment
+from htmresearch.frameworks.rl import DQN
+from htmresearch.frameworks.rl.openai_utils import create_atari_environment
 
 SEED = 42
 SPARSE = True
@@ -112,154 +109,6 @@ class SparseQ(nn.Module):
 
 
 
-class DQN(object):
-  """
-  Implements **DQN algorithm** described in https://arxiv.org/pdf/1312.5602.pdf
-  """
-
-
-  def __init__(self, environment, network,
-               eps_start=1.0, eps_end=0.01, eps_decay=0.9995,
-               learning_rate=0.0001, gamma=0.99, tau=1.0, target_update=1000,
-               batch_size=32, min_steps=10000, replay_size=10000):
-    """
-    :param environment: Open AI Gym environment
-    :param network: Neural neural network to use as a Q function approximator
-    :param eps_start: e-greedy exploration start epsilon
-    :param eps_end: e-greedy exploration end epsilon
-    :param eps_decay: e-greedy exploration decay
-    :param learning_rate: optimizer learning rate
-    :param gamma: future returns discount
-    :param tau: soft update smoothing coefficient. Use 1.0 for hard target update
-    :param target_update: target update interval
-    :param batch_size: batch size
-    :param min_steps: min number of experiences in replay buffer before learning
-    :param replay_size: replay memory size
-    """
-
-    self.env = environment
-
-    # e-greedy parameters
-    self.eps = eps_start
-    self.eps_end = eps_end
-    self.eps_decay = eps_decay
-
-    # Experience replay memory: e = (s, a, r, s', done)
-    self.replay = collections.deque(maxlen=replay_size)
-    self.min_steps = min_steps
-    self.batch_size = batch_size
-
-    # Initialize Local and Target networks
-    self.local = network
-    self.target = copy.deepcopy(self.local)
-    self.optimizer = optim.Adam(self.local.parameters(), lr=learning_rate)
-
-    # Optimizations
-    self.gamma = gamma
-    self.tau = tau
-    self.target_update = target_update
-    self.steps = 0
-
-
-  def select_action(self, state):
-    """
-    Select the best action for the given state using e-greedy exploration to
-    minimize overfitting
-
-    :return: tuple(action, value)
-    """
-    value = 0
-    if self.steps < self.min_steps:
-      action = self.env.action_space.sample()
-    else:
-      self.eps = max(self.eps_end, self.eps * self.eps_decay)
-      if random.random() < self.eps:
-        action = self.env.action_space.sample()
-      else:
-        self.local.eval()
-        with torch.no_grad():
-          state = torch.tensor(state, device=device, dtype=torch.float).unsqueeze(0)
-          Q = self.local(state)
-          value, action = torch.max(Q, 1)
-
-    return int(action), float(value)
-
-
-  def hard_target_update(self):
-    """
-    Update model parameters every 'target_update' time steps
-    See https://arxiv.org/abs/1312.5602
-    :param target_update: target update interval
-    """
-    if self.steps % self.target_update == 0:
-      self.target.load_state_dict(self.local.state_dict())
-
-
-  def soft_target_update(self):
-    """
-    Soft update model parameters:
-
-    .. math::
-      \\theta_target = \\tau \\times \\theta_local + (1 - \\tau) \\times \\theta_target ,
-      with \\tau \\ll 1
-
-    See https://arxiv.org/pdf/1509.02971.pdf
-    """
-    for target_param, local_param in zip(self.target.parameters(), self.local.parameters()):
-      target_param.data.copy_(self.tau * local_param.data + (1.0 - self.tau) * target_param.data)
-
-
-  def learn(self, state, action, reward, next_state, done):
-    """
-    Update replay memory and learn from a batch of random experiences sampled
-    from the replay buffer
-    :return: optimization loss if enough experiences are available, None otherwise
-    """
-    self.steps += 1
-    self.replay.append((state, action, reward, next_state, done))
-    if self.steps > self.min_steps and len(self.replay) > self.batch_size:
-      batch = random.sample(self.replay, self.batch_size)
-      return self.optimize(batch)
-
-    return None
-
-
-  def optimize(self, batch):
-    state, action, reward, next_state, done = map(
-      lambda x: torch.tensor(x, device=device, dtype=torch.float), zip(*batch))
-
-    # Get target values
-    self.target.eval()
-    with torch.no_grad():
-      target_next = self.target(next_state)
-      target_next = torch.max(target_next, dim=1, keepdim=True)[0]
-    Q_target = reward + self.gamma * target_next * (1 - done)
-
-    # Get actual values
-    self.local.train()
-    Q_local = self.local(state)
-    Q_local = Q_local.index_select(1, action.long())
-
-    # Compute Huber loss
-    loss = F.smooth_l1_loss(Q_local, Q_target)
-
-    self.optimizer.zero_grad()
-    loss.backward()
-    for param in self.local.parameters():
-      param.grad.data.clamp_(-1, 1)
-
-    self.optimizer.step()
-
-    # Update target model
-    if self.tau < 1:
-      self.soft_target_update()
-    else:
-      self.hard_target_update()
-
-    return loss.item()
-
-
-
 def run(env, model, learn, numSteps, render, resultsPath, checkpoint=None):
   """
   Main RL loop where the agent acts on the environment based on the states
@@ -282,7 +131,7 @@ def run(env, model, learn, numSteps, render, resultsPath, checkpoint=None):
     loss = 0
 
     # Train for numSteps
-    agent = DQN(env, model)
+    agent = DQN(env.action_space.n, model)
     state = env.reset()
     progess_bar = tqdm.trange(numSteps, desc="t")
     for t in progess_bar:
